+++ RTT/SEGGER_RTT_Conf.h
@@ -60,6 +60,10 @@ Revision: $Rev: 24316 $
 
 #include <rtt_cfg.h>
 
+#ifndef RTT_WRITE_SKIP_USE_ASM
+#define RTT_WRITE_SKIP_USE_ASM 0
+#endif
+
 #ifdef __IAR_SYSTEMS_ICC__
   #include <intrinsics.h>
 #endif